+++ packages/@expo/cli/src/start/server/metro/createServerRouteMiddleware.ts
@@ -115,9 +115,9 @@ export function createRouteHandlerMiddleware(
         }
       },
       async handleRouteError(error) {
-        const { ExpoError } = require('@expo/server') as typeof import('@expo/server');
-
-        if (ExpoError.isExpoError(error)) {
+        // NOTE(@kitten): temporarily vendored check instead of `ExpoError.isExpoError` to avoid API incompatibility for when
+        // we rename `@expo/server` if we ever hot-swap the implementation
+        if (error && typeof error === 'object' && error.name === 'ExpoError') {
           // TODO(@krystofwoldrich): Can we show code snippet of the handler?
           // NOTE(@krystofwoldrich): Removing stack since to avoid confusion. The error is not in the server code.
           delete error.stack;